+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Installation as a Clojure CLI [tool](https://clojure.org/reference/deps_and_cli#tool_install) by [pilosus](https://github.com/pilosus)
-- Report generation to a standard output by [pilosus](https://github.com/pilosus)
+- Report generation to standard output by [pilosus](https://github.com/pilosus)
 
 ## [0.2.1] - 2023-07-19